Fix type conversion error

+++ src/character.cpp
@@ -2393,7 +2393,8 @@ void Character::reset_stats()
 {
     // Bionic buffs
     if( has_active_bionic( bionic_id( "bio_hydraulics" ) ) ) {
-        int bonus = std::max( 20, std::floor( get_str_base() * .8 ) );
+        int bonus = std::max( 20, static_cast<int>( std::floor( get_str_base() * .8 ) ) );
+        
         mod_str_bonus( bonus );
     }
